:truck: rename my study info card & recruiting studies
