+++ src/pages/HomePage/MyStudyInfoCard/index.tsx
